Settled on a solution for TwoSum
Attempted to split twoSum into two functions but I ended up sticking with this solution. This seems to be what's commonly used.
The one difference among the regular solutions is whether to use the map {} object or the class Map. The latter allows the user to use class functions like .has(), etc.

I went with the object-solution, which seemed more intuitive.